@@ -16,6 +16,9 @@
     "setupFiles": [
       "../../../jest/environment.js"
     ],
+    "testPathIgnorePatterns": [
+      ".eslintrc.js"
+    ],
     "setupTestFrameworkScriptFile": "../../../node_modules/jest-enzyme/lib/index.js"
   },
   "devDependencies": {
@@ -32,12 +35,11 @@
     "babel-register": "^6.22.0",
     "css-loader": "^0.26.1",
     "enzyme": "^2.8.1",
-    "eslint": "^3.13.0",
-    "eslint-config-airbnb": "^14.0.0",
-    "eslint-config-airbnb-base": "^11.0.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^3.0.2",
-    "eslint-plugin-react": "^6.9.0",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.10.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.7.0",
     "espree": "^3.3.2",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
