@@ -390,7 +390,56 @@ slim_shims = {
         'EFO:0007098': ['skin of body'],
         'EFO:0007095': ['skin of body'],
         'EFO:0007097': ['skin of body'],
-        'EFO:0002779': ['skin of body']
+        'EFO:0002779': ['skin of body'],
+        'CL:0002399': ['blood', 'bodily fluid'],
+        'CL:0000236': ['blood', 'bodily fluid'],
+        'CL:0000084': ['blood', 'bodily fluid'],
+        'CL:0000625': ['blood', 'bodily fluid'],
+        'CL:0000624': ['blood', 'bodily fluid'],
+        'CL:0000897': ['blood', 'bodily fluid'],
+        'NTR:0003860': ['blood', 'bodily fluid'],
+        'CL:0000895': ['blood', 'bodily fluid'],
+        'CL:0000792': ['blood', 'bodily fluid'],
+        'CL:0000909': ['blood', 'bodily fluid'],
+        'CL:0000899': ['blood', 'bodily fluid'],
+        'CL:0000815': ['blood', 'bodily fluid'],
+        'CL:0000545': ['blood', 'bodily fluid'],
+        'CL:0000546': ['blood', 'bodily fluid'],
+        'CL:0000576': ['blood', 'bodily fluid'],
+        'CL:0001054': ['blood', 'bodily fluid'],
+        'CL:0000515': ['musculature of body'],
+        'CL:0000187': ['musculature of body'],
+        'NTR:0004646': ['skin of body', 'penis', 'connective tissue'],
+        'CL:0000905': ['blood', 'bodily fluid'],
+        'CL:0000037': ['blood', 'bodily fluid'],
+        'CL:0000192': ['musculature of body'],
+        'CL:0000746': ['musculature of body', 'heart'],
+        'CL:0000837': ['blood', 'bodily fluid'],
+        'NTR:0001407': ['brain'],
+        'UBERON:0001872': ['brain'],
+        'UBERON:0002021': ['brain'],
+        'EFO:0000222': ['blood', 'bodily fluid'],
+        'EFO:0002167': ['blood', 'bodily fluid'],
+        'EFO:0002357': ['blood', 'bodily fluid'],
+        'EFO:0007950': ['skin of body'],
+        'UBERON:0002114': ['small intestine', 'intestine'],
+        'UBERON:0002769': ['brain'],
+        'EFO:0002779': ['penis', 'skin of body', 'connective tissue'],
+        'EFO:0006270': ['lung', 'connective tissue'],
+        'NTR:0000475': ['skin of body', 'connective tissue'],
+        'NTR:0000476': ['skin of body', 'connective tissue'],
+        'NTR:0000477': ['skin of body', 'connective tissue'],
+        'NTR:0000478': ['skin of body', 'connective tissue'],
+        'NTR:0000479': ['skin of body', 'connective tissue'],
+        'NTR:0000480': ['mouth', 'connective tissue'],
+        'NTR:0000481': ['skin of body', 'connective tissue'],
+        'NTR:0000482': ['skin of body', 'connective tissue'],
+        'NTR:0000483': ['skin of body', 'connective tissue'],
+        'NTR:0000484': ['lung', 'connective tissue'],
+        'NTR:0000485': ['skin of body', 'connective tissue'],
+        'CL:0002328': ['lung', 'bronchus', 'epithelium'],
+        'CL:0002598': ['lung', 'bronchus'],
+        'CL:2000017': ['connective tissue', 'mouth']
     }
 
 }
