@@ -15,7 +15,7 @@ def reference_experiment_RNA_seq(testapp, lab, award):
     item = {
         'award': award['uuid'],
         'lab': lab['uuid'],
-        'status': 'started',
+        'status': 'in progress',
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
         'assay_term_name': 'RNA-seq'
@@ -29,7 +29,7 @@ def reference_experiment_RRBS(testapp, lab, award):
     item = {
         'award': award['uuid'],
         'lab': lab['uuid'],
-        'status': 'started',
+        'status': 'in progress',
         'assay_term_name': 'RRBS',
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829'
@@ -45,7 +45,7 @@ def reference_experiment_WGBS(testapp, lab, award):
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
         'lab': lab['uuid'],
-        'status': 'started',
+        'status': 'in progress',
         'assay_term_name': 'whole-genome shotgun bisulfite sequencing'
 
     }
@@ -59,7 +59,7 @@ def reference_experiment_chip_seq_control(testapp, lab, award, target_control):
         'lab': lab['uuid'],
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
-        'status': 'started',
+        'status': 'in progress',
         'assay_term_name': 'ChIP-seq',
         'target': target_control['uuid']
 
@@ -72,7 +72,7 @@ def reference_experiment_chip_seq_H3K27me3(testapp, lab, award, target_H3K27me3)
     item = {
         'award': award['uuid'],
         'lab': lab['uuid'],
-        'status': 'started',
+        'status': 'in progress',
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
         'assay_term_name': 'ChIP-seq',
@@ -89,7 +89,7 @@ def reference_experiment_chip_seq_H3K36me3(testapp, lab, award, target_H3K36me3)
         'lab': lab['uuid'],
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
-        'status': 'started',
+        'status': 'in progress',
         'assay_term_name': 'ChIP-seq',
         'target': target_H3K36me3['uuid']
 
@@ -104,7 +104,7 @@ def reference_experiment_chip_seq_H3K4me1(testapp, lab, award, target_H3K4me1):
         'lab': lab['uuid'],
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
-        'status': 'started',
+        'status': 'in progress',
         'assay_term_name': 'ChIP-seq',
         'target': target_H3K4me1['uuid']
 
@@ -117,7 +117,7 @@ def reference_experiment_chip_seq_H3K4me3(testapp, lab, award, target_H3K4me3):
     item = {
         'award': award['uuid'],
         'lab': lab['uuid'],
-        'status': 'started',
+        'status': 'in progress',
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
         'assay_term_name': 'ChIP-seq',
@@ -132,7 +132,7 @@ def reference_experiment_chip_seq_H3K27ac(testapp, lab, award, target_H3K27ac):
     item = {
         'award': award['uuid'],
         'lab': lab['uuid'],
-        'status': 'started',
+        'status': 'in progress',
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
         'assay_term_name': 'ChIP-seq',
@@ -147,7 +147,7 @@ def reference_experiment_chip_seq_H3K9me3(testapp, lab, award, target_H3K9me3):
     item = {
         'award': award['uuid'],
         'lab': lab['uuid'],
-        'status': 'started',
+        'status': 'in progress',
         'biosample_type': 'tissue',
         'biosample_term_id': 'UBERON:349829',
         'assay_term_name': 'ChIP-seq',
