@@ -172,5 +172,5 @@ class AnalysisStepRun(Item):
     embedded = [
         'analysis_step_version.analysis_step',
     ]
-    # Avoid using reverse links on this object as invalidating a virtual
+    # Avoid using reverse links on this object as invalidating a
     # step_run can cause thousands of objects to be reindexed.
